+++ Sistema-Scouts-Jose-Hernandez/src/main/java/ar/edu/utn/frc/sistemascoutsjosehernandez/dtos/events/EventFilterDTO.java
@@ -20,5 +20,5 @@ public class EventFilterDTO {
     private EventStatus status;
     private LocalDateTime dateFrom;
     private LocalDateTime dateTo;
-    private Long createdBy;
+    private Integer createdBy;
 }
